requirements.txt
flask>=2.3.0
groq>=0.5.0
python-dotenv>=1.0.0
